Move SkipIfVersion* funcs to the vault package
